Rename CurrentPeriodBalanceStatus component to PeriodDistributionSummary
