--- packages/kyodo-app/src/components/CurrentPeriodBalanceStatus.js
@@ -1,60 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { drizzleConnect } from 'drizzle-react';
-import styled from 'styled-components';
-import { getCurrentPeriodInfo } from '../reducers';
-
-const Wrapper = styled.div`
-  font-size: 24px;
-`;
-
-const ProgressBarWrapper = styled.div`
-  border: 1px solid #000000;
-  box-sizing: border-box;
-  border-radius: 20px;
-  height: 5px;
-  background: white;
-  overflow: hidden;
-  width: 170px;
-`;
-
-const ProgressBar = styled.div`
-  width: ${props => props.fraction * 100}%;
-  border-right: 1px solid #000000;
-  background: #bd6fd8;
-  height: 5px;
-`;
-
-const Percent = styled.div`
-  font-size: 11px;
-  margin-top: 4px;
-`;
-
-class CurrentPeriodBalanceStatus extends Component {
-  render() {
-    const { currentPeriod } = this.props;
-    const { currentBalance, initialBalance, periodTitle} = currentPeriod;
-    const distributed = initialBalance - currentBalance;
-    const fraction = (initialBalance > 0) ? distributed / initialBalance : 0;
-    const inPercent = parseFloat(fraction * 100).toFixed(2);
-    return (
-      <Wrapper>
-        {distributed} of {initialBalance} points distributed in {periodTitle}
-        <ProgressBarWrapper>
-          <ProgressBar fraction={fraction} />
-        </ProgressBarWrapper>
-        <Percent>{inPercent}%</Percent>
-      </Wrapper>
-    );
-  }
-}
-
-CurrentPeriodBalanceStatus.contextTypes = {
-  drizzle: PropTypes.object,
-};
-
-const mapStateToProps = state => ({
-  currentPeriod: getCurrentPeriodInfo(state),
-});
-
-export default drizzleConnect(CurrentPeriodBalanceStatus, mapStateToProps);
